refactor mcp-tutorial module-01 source code
