+++ gen-ai/mcp-tutorial/module-01/src/example-1/pyproject.toml
@@ -1,0 +1,10 @@
+[project]
+name = "weather"
+version = "0.1.0"
+description = "weather mcp server"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "httpx>=0.28.1",
+    "mcp[cli]>=1.6.0",
+]
