+++ gen-ai/mcp-tutorial/module-01/src/example-1/requirements.txt
@@ -1,0 +1,2 @@
+httpx>=0.28.1
+mcp[cli]>=1.6.0
