+++ gen-ai/mcp-tutorial/module-01/src/example-2/requirements.txt
@@ -1,0 +1,4 @@
+boto3>=1.38.3
+httpx>=0.28.1
+mcp[cli]>=1.6.0
+pillow>=11.2.1
